Refactor remote player as preparation for stepheight honoring
