+++ src/remoteplayer.h
@@ -1,0 +1,44 @@
+/*
+Minetest
+Copyright (C) 2010-2015 celeron55, Perttu Ahola <celeron55@gmail.com>
+This program is free software; you can redistribute it and/or modify
+it under the terms of the GNU Lesser General Public License as published by
+the Free Software Foundation; either version 2.1 of the License, or
+(at your option) any later version.
+This program is distributed in the hope that it will be useful,
+but WITHOUT ANY WARRANTY; without even the implied warranty of
+MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+GNU Lesser General Public License for more details.
+You should have received a copy of the GNU Lesser General Public License along
+with this program; if not, write to the Free Software Foundation, Inc.,
+51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA.
+*/
+
+#ifndef __REMOTEPLAYER_H__
+#define __REMOTEPLAYER_H__
+
+#include "player.h"
+#include "gamedef.h"
+
+/*
+	Player on the server
+*/
+class RemotePlayer : public Player
+{
+public:
+	RemotePlayer(IGameDef *gamedef, const char *name);
+
+	virtual ~RemotePlayer() {}
+
+	void save(std::string savedir);
+
+	PlayerSAO *getPlayerSAO()
+	{ return m_sao; }
+	void setPlayerSAO(PlayerSAO *sao)
+	{ m_sao = sao; }
+	void setPosition(const v3f &position);
+
+private:
+	PlayerSAO *m_sao;
+};
+#endif
