+++ src/server.cpp
@@ -62,6 +62,7 @@ with this program; if not, write to the Free Software Foundation, Inc.,
 #include "util/base64.h"
 #include "util/sha1.h"
 #include "util/hex.h"
+#include "remoteplayer.h"
 
 class ClientNotFoundException : public BaseException
 {
